See descriptions
AWS support removed from the models. It will be added later once I know
how to use it, this is because Heroku does `python manage.py
collectstatic --noinput` when deploying the app; though the static
collection can be disabled via Heroku CLI but thats an extra step.
Until I figure out how to handle this, the best thing is to put the
keys in `settings/base.py`